thoroughly kill all submodules
okay, that was getting thoroughly confusing with all of the different
places I could create commits from.

I'm afraid this all has to be one big megarepo to be manageable.
If a utility is ever good and stable enough to be worth using in
other projects, then I might let it "graduate" out of this repo.

I thought I removed the submodules earlier, but apparently I did it
the stupid way and had to do 'git rm --cached -r' to kill the zombies
I left behind.