more tracing; fix performance of polarization
I knew `polarization` would be inefficient when I wrote it, but
it went well above and beyond my expectations.  Apparently this is
what the program wasted nearly all of its time doing when I tried
to let this run overnight.

No *meaningful* code has changed (only diagnostics) so the results
I already computed are compatible with results from this.