Workspace? What workspace? Don't be silly.
So today the RLS got updated for me and things got bad.

Like, way bad.

Like, 8 hours of debugging, reinstallation, and minimizing
a code sample for a bug report only to discover at some point that
the last several commits you made while minimizing all actually
run fine if you do it in a certain way but that this doesn't
apply to some of the commits before that and nothing makes any
goddamn sense and you know what screw it this isn't a workspace
anymore bad