add array mapping and inv/det funcs
I finally reached a breaking point with the *_from_fn utilities.
They are the wrong abstraction for 90% of what I use them for.

The trouble is that vec_from_fn is generic in the output type,
when in reality I have used it almost exclusively for mapping
the elements of an existing array; this can lead to bugs if the
output size is inferred to be different from the input.

That said, the *reason* I've stuck with it for so long is because I
know full well that `map` requires much more design to be widely applicable.
`*_from_fn` gives you the ability to "zip-map" over multiple arrays for
free, as well as the ability to abstract over values and references.
For now, `map_arr` is limited to a single `[_; n]` by value, so there
will probably be a lot of dereferencing of borrowed Copy arrays in my code
until the next iterative improvement.
C'est la vie.

---

Also, I always meant for inv/det to be free functions.