phonopy file-based api
This replaces the phonopy-io command API with one that more easily
allows a directory to be preserved on demand and reused, and which
provides configuration more uniformly through builder interfaces.

This will enable the creation of smaller CLI stubs for performing
smaller parts of certain tasks.

The downside is that there are more intermediate states to the
computation that must be represented by distinct types, which
makes the implementation more complicated, and also results in
inconveniences such as e.g. Options returned from `gamma_eigensystem`.
(this is like the problem with clap's `value_of` for required args)