begin absorbing binary shims into rsp2_tasks
up until this point, rsp2_tasks has more or less been trying to
pretend that it is a "library of things a binary can use", when in
reality it was doing an awful lot of things that a library simply
has no business doing, in order to e.g. set up its logging and
to manage things like output directories.

My new vision for this crate is that it shall *only* expose main
functions for binaries, that it will contain *all* of the main
functions, and that its contents will basically be "whatever isn't
worth pulling out into another crate."

This also does the following things, which were partly what
motivated the redesign:

- Fixes #36
- Adds support for multiple configs