fix the 'static bounds on Partition
add a lifetime parameter to Partition, to eliminate 'static bounds.
In theory, this lets you partition a vector of references.
(in all honesty, it really only lets me stop complaining about
 unnecessary 'static bounds)

...what I WANTED to do was add an alternative trait for
cfg(feature = "beta") that has an associated type Unlabeled
(usually defined in impls using 'impl Trait'), so that we can
get rid of all the Boxes; but it turns out that impl Trait in
associated types is not supported yet.