rust 1.26: use existential types
*thud*

so, there's basically nothing for rsp2 to get out
of this feature at present because it can't be used
in traits yet, and even then, rsp2 has already been
forced by the hand of nature to be written in a way
that does not require existential types.

new code may benefit significantly from the ability
to return 'impl Fn', I hope.