hlist meta take 22, スタート!!
I think I've got it this time.

Structure is not the solution to anything. It has been the problem
all along.

My prior concern was that maybe I might accidentally rearrange
the coordinates without doing the same to metadata. ...yet, up
until how, I already HAVE had to worry about this, because I do
not store all metadata in the Structure.  (because for things
which are optional like Layers, it is far too unergonomic to have
the Options embedded in an AoS format, compared to having a single
`Option<Vec<Layer>>`).

Likewise, Structure was supposed to help with replicating data for
a supercell, but clearly I was not satisfied with that use case
considering that `SupercellToken::replicate` exists.  (Meanwhile,
the very existence of Structure requires SupercellToken to have this
tortured interface that tries to support both Coords and Structure!)

In past attempts to add hlists for metadata, I've found that some
really silly allowances had to be made; like making Coords generic
over its storage type to support "Structure references" that borrow
their coords and a subset of the metadata. But without a Structure to
contain it, `&Coords` can be passed around just fine.

So the aim of this branch isn't just to introduce HLists for metadata.

The aim is to kill Structure.