remove unused alternate Fn and IsNewtype
IsNewtype was added at some point thinking I would be working with
a type like `&Json<[T]>`.  I'm not sure what I was thinking.

IIRC, alternate::Fn was added when I tried to add typelevel options
to gamma_analysis, which never panned out.  There are a number of things
it *could* be used for, but nothing that it currently *is* used for.
Rip it out.