flesh out requirements for safety on IndexCast
I had a sudden "fridge horror" realization that, by transmuting the
keys of maps and sets, I am making a variety of assumptions about
the impls of safe standard library traits like Hash and Ord.
The added documentation tries to address this though I'm still not
sure if it is bulletproof.

(or idiotproof, at least, considering that I'm the only one using it)