Work on conditional deps
This makes use of the TOML-patching feature added to the crates script
in order to make some dependencies of some crates conditional. In some cases
they are unconditionally activated by crates higher up the dependency tree.

I was hoping this would finally let me expose a feature named "mpi", but...
unfortunately, it turns out that cargo's optional dependencies cause an awful
lot of problems! For each optional dependency I had to add a feature with a
"-support" suffix, because the implicit features from the optional deps themselves
cannot be connected in a scalable manner.

We'll need to wait on cargo's unstable 'namespaced-features' to remove the
"-support" suffixes.