add option to use rebo instead of rebo/omp
activate it with the following config:

potential:
  rebo:
    omp: false

This helps work around https://github.com/ExpHP/rsp2/issues/64

Notice that this change breaks existing config files that used
rebo, because it now serializes as a singleton mapping instead of
a string. What used to be 'rebo' should now be 'rebo: {}'.
(making it serialize as a string was in hindsight a bad idea)