restore build config with working allowed_failures
Despite the documentation providing no indication of this,
"name:" in allowed_failures works!

The overall change compared to HEAD~1 is just the replacement
of the "env: ..." line in allowed_failures with "- name: ...".
(notice that this also fixes it to be a sequence rather than a
mapping, which is why it did not work at all before)