fix terrible bug in bond graph
The fix for #68 was backwards, and so was the test.
This basically entirely broke the bond graph whenever an atom
exited the unit cell; the bonds for images would be there, but
the stored image diffs would be absurd (giving the vectors
large lengths once converted back to cartesian).
Simply terrible.

FracBonds now performs a sanity check on the bond lengths in
debug mode.