actually clip as advertised; also remove splines
The previous commit forgot to actually do the clipping
into (3,3,9).  When I tried to add this, I couldn't find
a way to do it that didn't feel like a terrible hack (we have
these 4x4x9 splines that are being clipped to 3x3x9).

So I ultimately decided to just remove the spline implementations
since they weren't actually being used, and the DRYness of their
implementation was the design constrant that was forcing us to
use a 4x4x9.