investigation into numerical error of G
I was curious why compute_paranoid was failing for seemingly
generous tolerances of 1e-9 on some terms. It turns out that,
if cos_ijk is near one of the knot values, then the error in
the numerical derivative may scale as O(h^3) rather than O(h^4)
due to the discontinuity in the fourth derivative.

In other words, the forces computed from G are fine. Rather, it
is the numerically-computed derivative we compare them to (our
"expected" value) that is inaccurate!