refactor acgsd output
I need to move the stop_condition logic outside of minimize for
us to be able to put conditions on atomic forces when relaxing
with lattice params.

This is a simple PoC "first step" towards that, by factoring out the
output logging function.

Also remove the bin reporting and drift_spec, neither of which has
proved helpful.