better natural comparison in sorted-diff
Use our own natural comparison scheme so that it
can be reflected in the diff output. (i.e. stop showing
-0.00000 and 0.00000 as differences)

I did this ages ago but somehow never it was never
commited to VCS it seems :/. Anyways it was something
I could easily re-add while on the road with no charger.