REBO: add cyclopropane tests
When LAMMPS computes the dihedral term, it specifically checks
that k != l.  That would be annoying for us to do, because we would
have to use the FracBond composition API in order to make sure that
different images of a site are still considered distinct.

IIRC it currently does not have this test because the dihedral term
naturally comes out to zero when k == l. I added these tests to confirm
that the current implementation (appears to) work.

Cyclopropane has a triangle of carbons, resulting in a quadruplet
where k == l. However, the T spline is zero in this structure, so
the term is short-circuited away.  Therefore, I also added a
structure that I've named "psychopropane" where three of the hydrogens
are missing.