split up do_post_relaxation_computations
This has been one of the most desparately-needed refactorings
since forever. (and it is required for the next binary I need to add!)

The only reason that the 'stop_after_dynmat' hack was embedded all
the way into 'do_post_relaxation_computations' was because formerly,
due to the option to use phonopy, the code to diagonalize the diagonal
matrix was highly coupled to the code that produced the dynamical matrix.

Without the option to use phonopy, the function now cleanly splits into
three functions; one that computes the dynamical matrix, one that saves
it to a file, and one that diagonalizes it. Getting rid of the confusing
Option<Iteration> argument was also a fantastic improvement.

The 'stop_after_dynmat' hack still exists, but it is *far* less
terrible now.