allow convert script to read stdin/stdout
DwimHooks was added because in the past there have been numerous
bugs due to forgetting to add an argument to the signature of a
callback.  tp.Callables can't have keyword arguments, so I needed
an abstract base class to get the linter to successfully check
the existence of a 'file' argument.