unfold: Add missing phases to the eigenvector
It turns out that Allen's method cannot be applied directly
to the phonon eigenvectors, because they are not bloch functions
with the appropriate wave vector. (they are always periodic, with
no phase difference between cells)

Rather, it is the displacements that have the right set of
mathematical properties.  Compared to the eigenvectors, the
displacements have an additional e^-i(q.r) factor that gives
them the desired properties.