truly correctly fix phases in unfold
This is it.  I am now 100% certain that all signs in this script
are correct.

The cause of confusion earlier was that the dynmat was generated
using phonopy's phase convention, while the unfold script was trying
to follow Allen's phase convention. For consistency, I decided it is
best to use phonopy's phase convention everywhere.

I also had to re-work through the explanation for translation_phases
to convince myself that it is still correct even now that the bloch
function includes phases at each site.  Working through this also made
it trivial to verify the new sign factor.