update lapacke and lapack-src
lapack-src turned out to be the one responsible for the
lack of multithreading on matisse.  Well, not lapack-src
itself (the only thing that changed from 0.1.4 to 0.3
were the versions of its dependencies).  The true culprit
was probably fixed in the openblas-src 0.5 -> 0.6 update,
but it's hard to get a good look at the changes since the
author doesn't use tags.