add back sorting, but a bit faster
So it turns out that without the sorting, some raman.json files
look completely different (the butterfly effect may be causing
different degenerate kets to be chosen?) and that seems nasty to
deal with.  I prefer for the rust potentials to produce output as
similar to LAMMPS as possible for ease of comparison.

With the sort the way it is written in this commit, we still at least
have something that's faster than it was prior to binning. (it reduces
the time spent on my largest structures from 7.5 seconds to 2.5 seconds)

I tried adding parallelism to `frac_bonds`, but it very quickly
hit the limit of Amdahl's law and couldn't go faster than 0.9 seconds.
That complicated the API quite a bit more (I don't want rsp2-structure
to require rayon, so I had to introduce a cargo feature for it, and a
Builder for FracBonds to allow for an optional `use_rayon` argument...).
I'd rather focus on finding ways to speed it up further without
embedding parallelism.