automatically generate bond graphs sometimes
This is going to come back to bite me but I don't have
much of a choice.  Automatically generates bond graphs
in the KC-Full implementations when one isn't available,
so that these potentials can be used on layers.yaml.

Strictly speaking it would be better if it always generated
them; I no longer think the global 'bond-radius' config
is a very good idea.