don't combine CH.airebo with pair_style rebo
LAMMPS now errors if you do this (breaking backwards compatibility
and etc....).  I'm not 100% convinced that this is even correct,
because these params are not the only hydrogen-related things that
changed between 2nd-gen REBO and AIREBO. LAMMPS still uses AIREBO's
forms for G(cos(theta)) and exp^lambda, so it's still mixing and
matching stuff no matter which file you use (it's just a question
of which mismatched pairing is *worse*).

But whatever.