make SupercellToken more clearly for arbitrary matrices
Even though SupercellToken has always been written with non-diagonal supercells in mind,
for some reason it didn't have a method to get the matrix, and also, the 'periods' method
was a bit too footgunny; there was, in fact, code in rsp2 that incorrectly used it by
assuming the supercell was diagonal.  This has been renamed to iteration_periods, and
a fallible 'as_diagonal' has been introduced.