add sign-error toggle for bond polarizability
A bug was identified by Mayada Fadel in that the bond polarizability
code has an errant `+` instead of a `-` somewhere which changes the
symmetry of one of the terms.  Not good.

For comparison purposes and/or reproducing old data, the old behavior
can be brought back with the settings YAML:

bond-polarizability:
  form: with-sign-error

There's a bunch of tiny infrastructure changes in this commit to allow
bond polarizability to read config.  Honestly it should have always done
so, and bond polarizability coefficients would ideally also be
configurable through this; but for now it's just to toggle the sign error.